requirements.txt
numpy==1.22.4
matplotlib==3.5.2
requests==2.27.1
Pillow==9.0.1